+++ app/Domains/Website/Controllers/WebOpdController.php
@@ -70,7 +70,7 @@ class WebOpdController extends BaseController
         $pdfContent = $result['dompdf']->output();
         return $this->response
             ->setContentType('application/pdf')
-            ->setHeader('Content-Disposition', 'attachment; filename="' . $result['filename'] . '"')
+            ->setHeader('Content-Disposition', 'inline; filename="' . $result['filename'] . '"')
             ->setBody($pdfContent);
     }